+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 
 ## v0.1.15
 
+
 - fixed: now setGroupMembers works the right way, sending group-update notifications to original group members
 
 ## v0.1.14